+++ src/main/java/com/ssafy/solvedpick/problem/domain/Problem.java
@@ -2,7 +2,7 @@ package com.ssafy.solvedpick.problem.domain;
 
 import com.ssafy.solvedpick.api.dto.ProblemData;
 import com.ssafy.solvedpick.api.dto.SolvedProblemsApiResponse;
-import com.ssafy.solvedpick.common.point.Point;
+import com.ssafy.solvedpick.common.utils.point.Point;
 import com.ssafy.solvedpick.members.domain.Member;
 
 import jakarta.persistence.Column;
